requirements.txt
RealtimeSTT
langchain
langchain-ollama
langgraph
pyperclip
python-dotenv
keyboard
streamlit
